change to unsinged long int

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -8,7 +8,7 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int diff, count = 0;
+	unsigned long int diff, count = 0;
 
 	diff = n ^ m;
 	while (diff > 0)
